+++ Changelog.md
@@ -4,6 +4,8 @@ Jason Sekhon(Captain), Brayden Traas(Mate), Eric Lin(Mate), Morris Arroyo(Mate)
 
 -------------------------
 
+* added flights model, controller and view
+
 * menubar and boostrap integration. Cleaned welcome page
 
 * Added Airport and Airport models